Remove trush.
